refactor(hosting): split nodejs/mod.rs into per-group files
nodejs/mod.rs held every nested group (app/job/deployment/source/
github/secrets) inline in one 1088-line file. app/ and source/ get
their own subdirectories (multiple substantial leaf commands each);
job/deployment/github/secrets stay flat files. mod.rs keeps only the
two bare top-level commands and the shared client/terminal-status
helpers.
