+++ rust/src/hosting/nodejs/app/get.rs
@@ -1,0 +1,32 @@
+use cli_engine::{CommandResult, CommandSpec, NextActionParam, RuntimeCommandSpec, Tier};
+
+use crate::hosting::nodejs::{AppIdArgs, client_err, make_client};
+use crate::next_action::next_action;
+use crate::scopes::HOSTING_APPS_READ as APPS_READ;
+
+pub(super) fn command() -> RuntimeCommandSpec {
+    RuntimeCommandSpec::new_typed_with_context::<AppIdArgs, _, _, _>(
+        CommandSpec::from_args::<AppIdArgs>("get", "Get a Node.js hosting application")
+            .with_long("Get details for one Node.js hosting application.")
+            .with_system("hosting")
+            .with_tier(Tier::Read)
+            .with_scopes(&[APPS_READ]),
+        |ctx, args: AppIdArgs| async move {
+            let app_id = args.app_id;
+            let client = make_client(&ctx, &[APPS_READ]).await?;
+            let data = client.get_app(&app_id).await.map_err(client_err)?;
+            Ok(CommandResult::new(data).with_next_actions(vec![
+                next_action(
+                    "hosting nodejs deployment list --app-id <app-id>",
+                    "List deployments for this application",
+                )
+                .with_param("app-id", NextActionParam::value(app_id.clone())),
+                next_action(
+                    "hosting nodejs status --app-id <app-id>",
+                    "Get application status",
+                )
+                .with_param("app-id", NextActionParam::value(app_id)),
+            ]))
+        },
+    )
+}
